refactor: null인 경우를 고려해 BigDecimal 형태로 반환(double은 null 허용 x)

+++ src/main/java/kspo/onfit/voucher/dto/VoucherResponseDto.java
@@ -21,8 +21,8 @@ public record VoucherResponseDto(
         Integer memberCount,
         Integer price,
 
-        double latitude,
-        double longitude,
+        BigDecimal latitude,
+        BigDecimal longitude,
 
         Long likeCnt,
         Boolean myLike
@@ -46,8 +46,8 @@ public record VoucherResponseDto(
                 voucher.getMemberCount(),
                 voucher.getPrice(),
 
-                voucher.getLat().doubleValue(),
-                voucher.getLng().doubleValue(),
+                voucher.getLat(),
+                voucher.getLng(),
 
                 likeCnt,
                 myLike
